dns remove whitelist
The whitelist and assembling the string to check it, are all meaningless
for the captive portal dns function.

It's actually a problem that if a client is using this as dns, not matching
the whitelist just doesn't reply, making the client slow down.

If a client is using this as DNS, it's because it has joined with us as the
AP, in that case direct EVERY dns request to our IP.

Then we don't need a whitelist nor even to parse the request, remove both.